upload  network optimization problem
This Matlab code reads network data file and builds  trees toward single or multiple gateways. The path length is limited by the number of hops.
This is the first step of solving the mathematical model in paper  "Network Coding Design For Reliable Wireless Sensor Networks" which contains two steps. The second step will be to choose the best nodes to perform network coding when failure scenario occurred.